TECh-2: Correct docker-compose setup.
