+++ MainApp/Client/Dockerfile
@@ -4,6 +4,6 @@ LABEL maintainer="saksofon997@gmail.com"
 
 WORKDIR /usr/src/Client 
 COPY . .
-# RUN ["npm", "install"]
+RUN ["npm", "install"]
 EXPOSE 4200
 CMD [ "npm", "start"]
