+++ MainApp/Servers/eureka/src/main/java/com/spring/eureka/EurekaApplication.java
@@ -2,8 +2,10 @@ package com.spring.eureka;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.netflix.eureka.server.EnableEurekaServer;
 
 @SpringBootApplication
+@EnableEurekaServer
 public class EurekaApplication {
 
 	public static void main(String[] args) {
