+++ MainApp/Servers/xmlws/Dockerfile
@@ -3,6 +3,8 @@ FROM maven:3.6.3-ibmjava-8-alpine AS backend
 LABEL maintainer="saksofon997@gmail.com"
 
 WORKDIR /usr/src/server
+COPY pom.xml /tmp/pom.xml
+RUN mvn -B -f /tmp/pom.xml -s /usr/share/maven/ref/settings-docker.xml dependency:resolve
 COPY . .
 RUN ["mvn", "package", "-DskipTests"]
