+++ MainApp/Servers/xmlws/src/main/java/com/java/xmlws/XmlwsApplication.java
@@ -2,8 +2,10 @@ package com.java.xmlws;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
 
 @SpringBootApplication
+@EnableEurekaClient
 public class XmlwsApplication {
 
 	public static void main(String[] args) {
