+++ MainApp/Servers/xmlws/src/main/java/com/java/xmlws/controller/TestController.java
@@ -1,0 +1,17 @@
+package com.java.xmlws.controller;
+
+import org.springframework.http.MediaType;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping(value = "api")
+public class TestController {
+
+    @GetMapping(path = "/test",
+            produces = MediaType.APPLICATION_JSON_VALUE)
+    public String getClinic() {
+        return "CAO LJUDI SAD SE RELOADUJE";
+    }
+}
