--- MainApp/Servers/xmlws/src/main/java/controller/TestController.java
@@ -1,18 +1,0 @@
-package controller;
-
-import org.springframework.http.MediaType;
-import org.springframework.web.bind.annotation.CrossOrigin;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping(value = "api")
-public class TestController {
-
-    @GetMapping(path = "/test",
-            produces = MediaType.APPLICATION_JSON_VALUE)
-    public String getClinic() {
-        return "caogfgMIHAJLO";
-    }
-}
