+++ MainApp/Servers/zuul/src/main/resources/application.properties
@@ -1,1 +1,4 @@
+ribbon.eureka.enabled=true
 
+hystrix.command.default.execution.isolation.thread.timeoutInMilliseconds=60000
+ribbon.ReadTimeout=60000
