+++ MainApp/Servers/zuul/src/main/resources/bootstrap.properties
@@ -1,0 +1,6 @@
+eureka.client.service-url.defaultZone=http://${REGISTRY_HOST:localhost}:${REGISTRY_PORT:8761}/eureka/
+eureka.client.service-url.healthcheck.enabled=true
+eureka.instance.prefer-ip-address=true
+
+server.port=8083
+spring.application.name=${SPRINGBOOT_APP_NAME:zuul}
