+++ docker-compose.yml
@@ -44,6 +44,8 @@ services:
       - eureka-serviceregistry
       - postgres
       - axon-server
+    volumes: 
+      - maven-repo:/root/.m2
 
   front:
     build: ./MainApp/Client
@@ -70,6 +72,8 @@ services:
       - postgres
       - axon-server
       - xmlws
+    volumes: 
+      - maven-repo2:/root/.m2
 
 networks:
   demo:
@@ -78,3 +82,5 @@ networks:
 volumes:
   database-data:
   front-storage:
+  maven-repo:
+  maven-repo2:
